route validator is created to create results

+++ src/db/migrations/20240301225318-add_subjective_result_table.ts
@@ -23,7 +23,7 @@ module.exports = {
         onDelete: 'CASCADE',
       },
       value: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.SMALLINT,
         allowNull: false,
       },
       phraseId: {
